Mod: `int16bytes_conv.js` module moved to `util/` - for organize modules
